+++ src/ai_karen_engine/__init__.py
@@ -1,11 +1,12 @@
 """Core runtime components for Kari."""
 
+
 def __getattr__(name):
     if name == "SLMPool":
         from ai_karen_engine.clients.slm_pool import SLMPool as _SLMPool
         return _SLMPool
     if name == "LLMOrchestrator":
-        from src.ai_karen_engine.llm_orchestrator import LLMOrchestrator as _LO
+        from ai_karen_engine.llm_orchestrator import LLMOrchestrator as _LO
         return _LO
     if name == "NightlyFineTuner":
         from ai_karen_engine.echocore.fine_tuner import NightlyFineTuner as _NT
@@ -17,13 +18,13 @@ def __getattr__(name):
         from ai_karen_engine.core.echo_core import EchoCore as _EC
         return _EC
     if name == "LNMClient":
-        from src.clients.transformers.lnm_client import LNMClient as _LC
+        from ai_karen_engine.clients.transformers.lnm_client import LNMClient as _LC
         return _LC
     if name == "BasicClassifier":
-        from src.clients.nlp.basic_classifier import BasicClassifier as _BC
+        from ai_karen_engine.clients.nlp.basic_classifier import BasicClassifier as _BC
         return _BC
     if name == "SpaCyClient":
-        from src.clients.nlp.spacy_client import SpaCyClient as _SC
+        from ai_karen_engine.clients.nlp.spacy_client import SpaCyClient as _SC
         return _SC
     raise AttributeError(name)
 
@@ -40,5 +41,3 @@ __all__ = [
     "BasicClassifier",
     "SpaCyClient",
 ]
-
-
