+++ src/ai_karen_engine/clients/nlp/__init__.py
@@ -1,0 +1,4 @@
+from ai_karen_engine.clients.nlp.basic_classifier import BasicClassifier
+from ai_karen_engine.clients.nlp.spacy_client import SpaCyClient
+
+__all__ = ["BasicClassifier", "SpaCyClient"]
