+++ src/ai_karen_engine/self_refactor/__init__.py
@@ -1,3 +1,4 @@
-from src.self_refactor import SelfRefactorEngine, PatchReport, SREScheduler
+from ai_karen_engine.self_refactor.engine import SelfRefactorEngine, PatchReport
+from ai_karen_engine.self_refactor.scheduler import SREScheduler
 
 __all__ = ["SelfRefactorEngine", "PatchReport", "SREScheduler"]
