--- src/clients/nlp/__init__.py
@@ -1,4 +1,0 @@
-from src.clients.nlp.basic_classifier import BasicClassifier
-from src.clients.nlp.spacy_client import SpaCyClient
-
-__all__ = ["BasicClassifier", "SpaCyClient"]
