+++ src/core/__init__.py
@@ -1,6 +1,6 @@
 """Core package initialization."""
 
 from ai_karen_engine.core.mesh_planner import MeshPlanner
-from src.core.gpu_training import gpu_optimized_train
+from ai_karen_engine.core.gpu_training import gpu_optimized_train
 
 __all__ = ["MeshPlanner", "gpu_optimized_train"]
