+++ src/core/autonomous_agent.py
@@ -1,9 +1,9 @@
 import time
 import uuid
 
-from src.core.prompt_router import PromptRouter
-from src.core.automation_manager import AutomationManager
-from src.core.workflow_engine_client import WorkflowEngineClient
+from ai_karen_engine.core.prompt_router import PromptRouter
+from ai_karen_engine.core.automation_manager import AutomationManager
+from ai_karen_engine.core.workflow_engine_client import WorkflowEngineClient
 from integrations.llm_utils import LLMUtils
