+++ src/core/plugin_router.py
@@ -83,7 +83,7 @@ class PluginRouter:
                 continue
 
             try:
-                module = importlib.import_module(f"src.plugins.{name}.handler")
+                module = importlib.import_module(f"ai_karen_engine.plugins.{name}.handler")
             except ModuleNotFoundError:
                 try:
                     spec = importlib.util.spec_from_file_location(
@@ -104,7 +104,7 @@ class PluginRouter:
             if os.path.exists(ui_path) and (manifest.get("trusted_ui") or advanced):
                 try:
                     spec = importlib.util.spec_from_file_location(
-                        f"src.plugins.{name}.ui", ui_path
+                        f"ai_karen_engine.plugins.{name}.ui", ui_path
                     )
                     ui_module = importlib.util.module_from_spec(spec)
                     assert spec.loader
