+++ src/self_refactor/__init__.py
@@ -1,5 +1,4 @@
-from src.self_refactor.engine import SelfRefactorEngine, PatchReport
-from src.self_refactor.scheduler import SREScheduler
+from ai_karen_engine.self_refactor.engine import SelfRefactorEngine, PatchReport
+from ai_karen_engine.self_refactor.scheduler import SREScheduler
 
 __all__ = ["SelfRefactorEngine", "PatchReport", "SREScheduler"]
-
