+++ tests/test_self_heal.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from src.self_refactor import SelfRefactorEngine
+from ai_karen_engine.self_refactor import SelfRefactorEngine
 
 
 class DummyLLM:
