+++ tests/test_self_refactor.py
@@ -1,5 +1,5 @@
 import pathlib
-from src.self_refactor import SelfRefactorEngine
+from ai_karen_engine.self_refactor import SelfRefactorEngine
 
 
 class DummyLLM:
