+++ tests/test_self_refactor_loop.py
@@ -1,4 +1,4 @@
-from src.self_refactor import SelfRefactorEngine
+from ai_karen_engine.self_refactor import SelfRefactorEngine
 
 
 class DummyLLM:
