+++ tests/test_sre_scheduler.py
@@ -1,6 +1,6 @@
 import os
 
-from src.self_refactor import SREScheduler, SelfRefactorEngine
+from ai_karen_engine.self_refactor import SREScheduler, SelfRefactorEngine
 
 
 class DummyEngine(SelfRefactorEngine):
